+++ lib/src/in_app_webview.dart
@@ -24,6 +24,7 @@ const javaScriptHandlerForbiddenNames = [
   "onAjaxProgress",
   "shouldInterceptFetchRequest",
   "onPrint",
+  "flutterInAppWebViewDropDownWorkaroud"
 ];
 
 ///InAppWebView Widget class.